work on #1 and stopping the Ad hoc that has been
There is no real use to put server on everything in the server base.
The structure is now formed in a way that assumes we are under the
serverbase.  Data is being pulled out of anywhere that a version is, ref
ex:.  the nginx version data is pulled to the web pillar and files were
adjusted to match the change.  Some older unused data referencing the
proof of concept was as removed.