switch in provision execution
Per a chat on the IRC #salt I found an area where the base env wouldn't
be run first.  To get around the issue, regardless of the root issues
getting resolved, we will be calling each env highstate to ensure order.
Also moved the bootstrap in to the state file area to reduce the need to
mount another folder for speed concerns