tries to simplify by allowing a json config file
This is simple idea, we just have a json file in the root of the local
with the vagrant file.  What happens is we then can set up the whole
local project with out needed for the funky reads and rights.  Well want
to pull in json files from the app though so we maintain the split
between the server and it's apps